requirements.txt
# Core dependencies
numpy>=1.22.0
opencv-python>=4.5.5.64
pillow>=9.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
jsonschema>=4.17.3
mss>=6.1.0
pynput>=1.7.6
pyobjc-core>=8.5.1; sys_platform == 'darwin'  # macOS only
pyobjc-framework-Quartz>=8.5.1; sys_platform == 'darwin'  # macOS only

# API and interface
fastapi>=0.85.0
uvicorn>=0.18.3

# Machine learning
torch>=1.11.0
torchvision>=0.12.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.8.0
isort>=5.10.1
mypy>=0.971
pylint>=2.15.3
